% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCo2.R
\name{CalculateCo2}
\alias{CalculateCo2}
\title{Calculate co2}
\usage{
CalculateCo2(pop, gdp, enInt, carbInt)
}
\arguments{
\item{pop}{Population size (in millions)}

\item{gdp}{GDP per capita (in 1000$/person)}

\item{entInt}{Energy Intensity (in Gigajoule/$1000GDP)}

\item{carnInt}{Carbon Intensity (in tonnes CO2/Gigajoule)}
}
\value{
result of co2 = pop * gdp * enInt * carbInt
}
\description{
The Kaya Identity is an equation that expresses yearly CO2 emissions as a product of four factors.
}
\examples{
CalculateCo2(82.4, 44, 5, 0.05)

}
